+++ core/model/auth.go
@@ -313,6 +313,7 @@ type ServiceRegistration struct {
 	Description  string                 `json:"description" bson:"description"`
 	InfoURL      string                 `json:"info_url" bson:"info_url"`
 	LogoURL      string                 `json:"logo_url" bson:"logo_url"`
+	VersionPath  string                 `json:"version_path" bson:"version_path"`
 	Scopes       []ServiceScope         `json:"scopes" bson:"scopes"`
 	AuthEndpoint string                 `json:"auth_endpoint" bson:"auth_endpoint"`
 	FirstParty   bool                   `json:"first_party" bson:"first_party"`